Complete integration of CUDA with build system
Until now, the configure script internally checked for presence of CUDA and compiled cuda modules if it detected CUDA. The user wasn't informed about whether CUDA was detected or if the CUDA modules will be compiled.

With this commit, the configure script prints a warning whenever CUDA is not detected and provides possible fixes for it.