Build Without Cuda option
Until now, the configure script checked for presence of cuda runtime and automatically
linked the library with cuda to allow use of GPU kernels. However users who do not want
to use the cuda funcitonality, even if they had a compatible gpu, did not have an
option to build the library without cuda.

With this commit, the configure script is modified to allow an extra argument 'without-cuda'.
When this option is passed to configure script, it does not look for presence of cuda runtime
and builds the library without cuda.
Note that the default mode(when no option is passed) is still to build with cuda.